requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.1
youtube-transcript-api==0.6.1
jinja2==3.1.2
python-multipart==0.0.6